adding cli package
