+++ cypher-shell/src/main/java/org/neo4j/shell/cli/InteractiveShellRunner.java
@@ -1,10 +1,13 @@
-package org.neo4j.shell;
+package org.neo4j.shell.cli;
 
 import jline.console.ConsoleReader;
 import jline.console.history.FileHistory;
 import jline.console.history.History;
 import jline.console.history.MemoryHistory;
 import org.neo4j.driver.v1.exceptions.ClientException;
+import org.neo4j.shell.BoltHelper;
+import org.neo4j.shell.Shell;
+import org.neo4j.shell.ShellRunner;
 import org.neo4j.shell.exception.CommandException;
 import org.neo4j.shell.exception.ExitException;
