+++ cypher-shell/src/main/java/org/neo4j/shell/cli/StringShellRunner.java
@@ -1,8 +1,9 @@
-package org.neo4j.shell;
+package org.neo4j.shell.cli;
 
 import jline.console.history.History;
 import org.neo4j.driver.v1.exceptions.ClientException;
-import org.neo4j.shell.cli.CliArgHelper;
+import org.neo4j.shell.Shell;
+import org.neo4j.shell.ShellRunner;
 import org.neo4j.shell.exception.CommandException;
 import org.neo4j.shell.exception.ExitException;
