+++ cypher-shell/src/test/java/org/neo4j/shell/StreamShell.java
@@ -1,6 +1,7 @@
 package org.neo4j.shell;
 
 import org.neo4j.driver.v1.exceptions.ClientException;
+import org.neo4j.shell.TestShell;
 import org.neo4j.shell.exception.CommandException;
 import org.neo4j.shell.exception.ExitException;
 
@@ -10,11 +11,11 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 
-class StreamShell extends TestShell {
+public class StreamShell extends TestShell {
     private final ByteArrayOutputStream errStream;
     private final ByteArrayOutputStream outStream;
 
-    StreamShell(@Nonnull final String input) {
+    public StreamShell(@Nonnull final String input) {
         in = new ByteArrayInputStream(input.getBytes());
         errStream = new ByteArrayOutputStream();
         err = new PrintStream(errStream);
