+++ cypher-shell/src/test/java/org/neo4j/shell/cli/AddressArgPatternTest.java
@@ -1,4 +1,4 @@
-package org.neo4j.shell.commands;
+package org.neo4j.shell.cli;
 
 import org.junit.Test;
 import org.neo4j.shell.cli.CliArgHelper;
