+++ cypher-shell/src/test/java/org/neo4j/shell/cli/InteractiveShellRunnerTest.java
@@ -1,6 +1,7 @@
-package org.neo4j.shell;
+package org.neo4j.shell.cli;
 
 import org.junit.Test;
+import org.neo4j.shell.StreamShell;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
