+++ cypher-shell/src/test/java/org/neo4j/shell/cli/NonInteractiveShellRunnerTest.java
@@ -1,8 +1,8 @@
-package org.neo4j.shell;
+package org.neo4j.shell.cli;
 
 import org.junit.Test;
 import org.neo4j.driver.v1.exceptions.ClientException;
-import org.neo4j.shell.cli.CliArgHelper;
+import org.neo4j.shell.StreamShell;
 import org.neo4j.shell.exception.ExitException;
 
 import static junit.framework.TestCase.fail;
